requirements.txt
# Core Python dependencies for Prismatron LED Display Software
# This file lists all required packages with version constraints

# Core numerical and scientific computing
numpy>=1.21.0,<3.0.0
scipy>=1.7.0,<2.0.0

# Computer vision and image processing
opencv-python>=4.5.0,<5.0.0
Pillow>=8.0.0,<12.0.0

# Deep learning and GPU acceleration
torch>=1.9.0,<3.0.0
torchvision>=0.10.0,<1.0.0

# Web framework and API
fastapi>=0.68.0,<1.0.0
uvicorn[standard]>=0.15.0,<1.0.0
websockets>=10.0,<12.0
python-multipart>=0.0.5,<1.0.0

# Video processing (FFmpeg Python bindings)
ffmpeg-python>=0.2.0,<1.0.0

# Data serialization and configuration
pydantic>=1.8.0,<3.0.0
PyYAML>=6.0,<7.0
toml>=0.10.0,<1.0.0

# Network and serial communication  
pyserial>=3.4,<4.0
requests>=2.25.0,<3.0.0

# Mathematical and symbolic computing
sympy>=1.8,<2.0
matplotlib>=3.3.0,<4.0.0

# Graph and network analysis
networkx>=2.6,<4.0

# Template engine for web interface
jinja2>=3.0.0,<4.0.0

# Process and system utilities
psutil>=5.8.0,<6.0.0

# Development and testing
pytest>=6.0.0,<8.0.0
pytest-asyncio>=0.15.0,<1.0.0
pytest-cov>=2.12.0,<5.0.0

# Code quality and formatting
black>=21.0.0,<24.0.0
flake8>=3.9.0,<7.0.0
mypy>=0.910,<2.0.0

# Documentation
mkdocs>=1.2.0,<2.0.0
mkdocs-material>=7.0.0,<10.0.0

# Logging and monitoring
structlog>=21.1.0,<24.0.0

# Date and time utilities
python-dateutil>=2.8.0,<3.0.0

# Job scheduling and parallel processing
joblib>=1.0.0,<2.0.0

# Mathematical optimization
scikit-learn>=1.0.0,<2.0.0

# Additional utilities
tqdm>=4.60.0,<5.0.0
click>=8.0.0,<9.0.0

# Security and encryption
cryptography>=3.0.0,<42.0.0

# Optional dependencies for enhanced functionality
# These may not be available on all systems but provide additional features

# Hardware acceleration (NVIDIA only)
# nvidia-ml-py3>=7.352.0,<12.0.0

# Advanced image processing
# scikit-image>=0.18.0,<1.0.0

# Additional video codecs
# imageio-ffmpeg>=0.4.0,<1.0.0

# Performance profiling
# line_profiler>=3.0.0,<5.0.0
# memory_profiler>=0.58.0,<1.0.0